requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.7.0
cachecontrol==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.1
dataclasses-json==0.6.3
distro==1.9.0
exceptiongroup==1.2.0
firebase-admin==6.4.0
flask==3.0.1
Flask-Cors==4.0.0
frozenlist==1.4.1
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.16
langchain-openai==0.0.5
langsmith==0.0.83
llamaapi==0.1.36
MarkupSafe==2.1.4
marshmallow==3.20.2
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.3
openai==1.10.0
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
PyJWT==2.8.0
pyparsing==3.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rsa==4.9
sniffio==1.3.0
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
typing-extensions==4.9.0
typing-inspect==0.9.0
uritemplate==4.1.1
urllib3==2.1.0
werkzeug==3.0.1
yarl==1.9.4
zipp==3.17.0
